Fixed Private Timeline Bug
The post box was missing a logged in check on the private timeline.
Meaning logged-out users could attempt to post from the private timeline. (unsuccessfully)
This could in some cases create an unnamed user that any person could post from, if they were logged out, and on the private timeline.