gitignore update for annoying database
SO, we need a database in the bin files for our integration tests to work.
But as we no-longer hard-copy a database on build it does not work.

As a "quick" fix, I will simply place a DB in the bin folder and exclude it from gitignore.

Hopefully it works.